Fluence NFTs Collection Proposal
This is my NFTs proposal, using vector Illustrator artwork of my own (except the Fluence logo mark, of course). The elements can be interchangeable to generate a big collection with some rarity elements, like the fire circle or the stellar explosion above the logo. Background can also be colored, black and white, silver or golden, mixing the attributes to create different rarity levels (I can deliver vector layered artwork or in Photoshop layers as it's preferred). The various circles around the main circle  logo focus try to show the different possibilities of web 3.0 in a matching way. Any input is very much appreciated. All the best to your bounty and activity, artespraticas